+++ CHANGES.md
@@ -6,6 +6,8 @@
 
   + Remove support for OCaml 4.06 and 4.07, minimal version requirement bumped to OCaml 4.08 (#1549, @gpetiot)
 
+  + Remove the 'extension-sugar' option, deprecated since 0.14.0 (#1588, @gpetiot)
+
 #### Bug fixes
 
   + Fix parsing of invalid file wrt original source handling (#1542, @hhugo)
@@ -34,7 +36,7 @@
   + Add buffer filename in the logs when applying ocamlformat (#1557, @dannywillems)
 
   + Improve comment position in pattern collection (#1576, @gpetiot)
-  
+
   + Consistent positioning of lambda return type annotations when no-break-infix-before-func and pre/post extensions (#1581, @gpetiot)
 
 ### 0.16.0 (2020-11-16)